cambios decimales

+++ WAConectorEcommerce/WATickets/Models/ControlProductosViewModel.cs
@@ -10,7 +10,7 @@ namespace WATickets.Models
         public int idProducto { get; set; }
         public string Item { get; set; }
         public string ItemName { get; set; }
-        public int CantidadUsado { get; set; }
+        public decimal CantidadUsado { get; set; }
          
     }
 }
